extracted files templates to file-upload.js
